Ratingroute almost completed(delete method got an error :(

+++ backend/routes/RatingRoute.js
@@ -1,18 +1,22 @@
 const router=require("express").Router();
 const Rating=require("../models/Rating");
 const Course=require("../models/Course");
+const {verifyToken,verifyAutherziationandToken,verifyAdminandToken}=require('../verify');
 //post a rating of a course
-router.post('/newrating/:courseid',async(req,res)=>{
-    let avgRating;
-    if(req.body) avgRating=req.body.expectations+req.body.instructor+req.body.duration+req.body.skillup;
-    avgRating/=4;
-    const rating=new Rating({...req.body,courseId:rea.params.courseid,avg:avgRating});
+router.post('/newrating/:courseid',verifyToken,async(req,res)=>{
     try{
+      const checkuser=await Rating.findOne({userId:req.user.id});
+      if(!checkuser){
+      let avgRating;
+      if(req.body) avgRating=req.body.expectations+req.body.instructor+req.body.duration+req.body.skillup;
+      avgRating/=4;
+      const rating=new Rating({...req.body,courseId:req.params.courseid,avg:avgRating,userId:req.user.id});
      const newrating=await rating.save();
      const course=await Course.findById(req.params.courseid);
      let newcourserating=(course.rating+newrating.avg)/(course.studentno+1);
      await Course.findByIdAndUpdate(req.params.courseid,{$set:{"studentno":course.studentno+1,"rating":newcourserating}},{new:true})
      res.status(200).json(newrating);
+     }else res.json("Already rated"); 
    }catch(err){
     console.log(err);
     res.json(err);
@@ -31,7 +35,7 @@ router.get('/get/courserating/:courseid',async(req,res)=>{
 // getting all ratings by userid
 router.get('/get/userrating/:userid',async(req,res)=>{
   try{  
-    const ratings=await Rating.find({userId:req.params.courseid});
+    const ratings=await Rating.find({userId:req.params.userid});
      res.status(200).send(ratings);
   }catch(err){
    console.log(err);
@@ -39,9 +43,10 @@ router.get('/get/userrating/:userid',async(req,res)=>{
   }
 })
 // updating a rating
-router.put('/update/rating/:ratingid',async(req,res)=>{
+router.put('/update/rating/:ratingid',verifyToken,async(req,res)=>{
   try{
      const oldrating=await Rating.findById(req.params.ratingid);
+     if(oldrating.userId===req.user.id){
      const course=await Course.findById(oldrating.courseId);
      let newrating=(course.rating*course.studentno)-(oldrating.avg);
      const rating=await Rating.findByIdAndUpdate(req.params.ratingid,{$set:req.body},{new:true});
@@ -51,22 +56,25 @@ router.put('/update/rating/:ratingid',async(req,res)=>{
      newrating+=avgRating;
      newrating/=course.studentno;
      const newcourserating=await Course.findByIdAndUpdate(updatedrating.courseId,{$set:{"rating":newrating}},{new:true});
-     res.status(200).send(updatedrating,newcourserating);
+      res.status(200).json({updatedrating,newcourserating});
+     }else req.status(401).send("unautherized");
   }catch(err){
    console.log(err);
    res.json(err);
   }
 })
 //deleting a rating
-router.delete('/delete/rating/:ratingid',async(req,res)=>{
+router.delete('/delete/rating/:ratingid',verifyToken,async(req,res)=>{
   try{
      const oldrating=await Rating.findById(req.params.ratingid);
+     if(oldrating.userId===req.user.id){
      const course=await Course.findById(oldrating.courseId);
      let newrating=(course.rating*course.studentno)-(oldrating.avg);
      newrating/=(course.studentno-1);
      await Course.findByIdAndUpdate(oldrating.courseId,{$set:{"rating":newrating}},{new:true});
      await Rating.findByIdAndDelete(req.params.ratingid);
      res.status(200).send("This teacher has been deleted");
+     }else res.status(401).json("Unautherized");
   }catch(err){
    console.log(err);
    res.json(err);
